+++ src/spectrum.cpp
@@ -5,7 +5,7 @@ Spectrum::Spectrum(size_t num_buckets_)
 {
 }
 
-std::pair<float, float> Spectrum::TotalHz() const
+std::pair<double, double> Spectrum::TotalHz() const
 {
     return std::make_pair(HzRange(0).first, HzRange(NumBuckets() - 1).second);
 }
